% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_grid.R
\name{define_grid}
\alias{define_grid}
\title{Define the grid}
\usage{
define_grid(model_spec, predictor_data, grid_type, ...)
}
\arguments{
\item{model_spec}{A model specification}

\item{predictor_data}{The predictor data used to finalize the grid}

\item{grid_type}{The type of grid}
}
\value{

}
\description{
Define the grid
}
