% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_model.R
\name{define_model}
\alias{define_model}
\title{Define the model}
\usage{
define_model(model, engine, mode, ...)
}
\arguments{
\item{model}{The model to be fit}

\item{engine}{The underlying engine}

\item{mode}{Classification or regression?}
}
\value{

}
\description{
Define the model
}
