% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_wflow.R
\name{define_wflow}
\alias{define_wflow}
\title{Define the workflow}
\usage{
define_wflow(recipe, model, ...)
}
\arguments{
\item{recipe}{A trained recipe}

\item{model}{A model specification}
}
\value{

}
\description{
Define the workflow
}
