% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Get the data from the UCI repository}
\usage{
get_data()
}
\value{
A list with the training set (data_train) and the testing set (data_test)
}
\description{
Get the data from the UCI repository
}
