% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preprocess}
\alias{preprocess}
\title{Preprocess the data for training}
\usage{
preprocess(training_set)
}
\arguments{
\item{training_set}{The training data}
}
\value{

}
\description{
Preprocess the data for training
}
