% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_lvl.R
\name{show_lvl}
\alias{show_lvl}
\title{Print the levels of the target variable}
\usage{
show_lvl(model_fit)
}
\arguments{
\item{model_fit}{A trained workflow}
}
\description{
Print the levels of the target variable
}
