% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_grid.R
\name{define_grid}
\alias{define_grid}
\title{Define the grid}
\usage{
define_grid(model, predictor_data, grid, ...)
}
\arguments{
\item{model_spec}{A model specification}

\item{grid_size}{The size of the grid}

\item{finalize}{Should the parameters be finalized (mtry for tree based model for instance)}
}
\value{

}
\description{
Define the grid
}
